Переместил типы в папку models
