+++ src/components/app/app.jsx
@@ -3,11 +3,11 @@ import Main from '../main/main';
 import PlaceDetails from '../place-details/place-details';
 import SignIn from '../sign-in/sign-in';
 import {arrayOf, func, shape, bool} from 'prop-types';
-import {placeType, cityType} from '../../mocks/offers';
+import {placeType, cityType} from '../../models/index';
 import {connect} from 'react-redux';
 import {Operation} from '../../reducers/index';
 import {ActionCreator} from '../../reducers/application/application';
-import {variantType} from '../../mocks/sort-variations';
+import {variantType} from '../../models/index';
 import {userParamsType} from '../../models/index';
 
 const sortPlaces = (places, sortVariant) => {
