+++ src/mocks/sort-variations.js
@@ -1,10 +1,3 @@
-import {shape, number, string} from 'prop-types';
-
-export const variantType = shape({
-  id: number,
-  name: string
-});
-
 export const variations = [
   {
     id: 0,
