+++ src/models/index.js
@@ -1,1 +1,3 @@
 export * from './user-params';
+export * from './offer';
+export * from './sort-variation';
