+++ src/models/sort-variation.js
@@ -1,0 +1,6 @@
+import {shape, number, string} from 'prop-types';
+
+export const variantType = shape({
+  id: number,
+  name: string
+});
